find_h5_file now works and a module netcdf_groundwater has been made
