+++ CHANGES.rst
@@ -5,6 +5,10 @@ threedi-qgis-plugin changelog
 0.16 (unreleased)
 -----------------
 
+- Add module base.netcdf_groundwater (relocated from base.DummyDataSourse)
+
+- Fix find_h5_file
+
 - Add (temporary) ad-hoc implementations of get_timeseries and
   get_values_by_timestemp_nr.
